Fix fucking bastard long to double conversion - adrien/anthony

+++ src/main/java/scheduling/Scheduler.java
@@ -104,7 +104,7 @@ public interface Scheduler {
         /**
          * Duration in ms of the reconfiguration.
          */
-        public long duration;
+        public double duration;
 
     }
